refactor (remove unused code)

+++ src/main/java/com/impllife/xlsx/service/ExcelService.java
@@ -1,37 +1,35 @@
 package com.impllife.xlsx.service;
 
 import com.impllife.xlsx.Const;
-import com.impllife.xlsx.data.Stat;
-import com.impllife.xlsx.data.StatSrt;
+import com.impllife.xlsx.data.StatByString;
 import com.impllife.xlsx.data.Transaction;
 import com.impllife.xlsx.data.map.ColumnDefinition;
 import com.impllife.xlsx.service.map.JSONLoader;
-import org.apache.poi.ss.usermodel.*;
+import org.apache.poi.ss.usermodel.Cell;
+import org.apache.poi.ss.usermodel.Row;
+import org.apache.poi.ss.usermodel.Sheet;
+import org.apache.poi.ss.usermodel.Workbook;
 import org.apache.poi.ss.util.CellRangeAddress;
 
 import java.io.File;
-import java.io.IOException;
 import java.math.BigDecimal;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.*;
 import java.util.stream.Collectors;
 
-import static com.impllife.xlsx.service.util.DateUtil.*;
+import static com.impllife.xlsx.service.util.DateUtil.concatDateAndTime;
+import static com.impllife.xlsx.service.util.DateUtil.getCurrentDateTime;
 import static com.impllife.xlsx.service.util.WorkbookUtil.*;
 import static org.apache.poi.ss.usermodel.Row.MissingCellPolicy.CREATE_NULL_AS_BLANK;
 
 public class ExcelService {
-
-    private final static List<String> ignore = new ArrayList<>();
-    static {
-        ignore.add("Зі своєї картки 51**22");
-//        ignore.add("Чудновська Вікторія Леонідівна");
-    }
     private final Map<String, List<ColumnDefinition>> columnDefinitionsMap = new HashMap<>();
+
     public ExcelService() {
         readMappings();
     }
+
     public void readMappings() {
         columnDefinitionsMap.clear();
         columnDefinitionsMap.put("mono", JSONLoader.loadColumnDefinitions(Const.MAP_DIR + "mono_input_excel_mappings.json"));
@@ -75,22 +73,10 @@ public class ExcelService {
         Workbook template = read(Const.TEMPLATES_DIR + "template.xlsx");
         String resultFileName = Const.RESULT_DIR + "stat_" + getCurrentDateTime() + ".xlsx";
         File resFile = new File(resultFileName);
-        if (resFile.exists()) {
-            resFile.delete();
-        } else {
-            try {
-                resFile.createNewFile();
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
+        if (resFile.exists()) resFile.delete();
 
         List<Transaction> transactions = readAndSortFilesTransactions();
         Map<String, List<Transaction>> groupByMonth = groupByMonth(fillEmptyDays(transactions));
-
-//        Sheet monthStat = cloneSheet(template, "T1", "Months");
-//        putStat(monthStat, monthStat);
-//        Sheet allStat = cloneSheet(template, "T3", "Months");
 
         SimpleDateFormat dateFormat = new SimpleDateFormat("MM.yyyy");
         List<Map.Entry<String, List<Transaction>>> sortedGroupByMonth = groupByMonth.entrySet().stream().sorted((e1, e2) -> {
@@ -107,18 +93,19 @@ public class ExcelService {
         for (Map.Entry<String, List<Transaction>> entry : sortedGroupByMonth) {
             Sheet monthStat = cloneSheet(template, "T2", entry.getKey());
             List<Transaction> transactionList = entry.getValue();
-            List<StatSrt> monthStatistic = transactionList.stream()
-                .collect(Collectors.groupingBy(Transaction::getDscr))
-                .entrySet().stream()
+            List<StatByString> monthStatistic = transactionList.stream()
+                .collect(Collectors.groupingBy(Transaction::getDscr)).entrySet().stream()
                 .map(e -> {
-                    StatSrt stat = new StatSrt();
+                    StatByString stat = new StatByString();
                     stat.setStr(e.getKey());
-                    BigDecimal sum = e.getValue().stream().map(Transaction::getSum).reduce(BigDecimal.ZERO, BigDecimal::add);
+                    BigDecimal sum = e.getValue().stream()
+                        .map(Transaction::getSum)
+                        .reduce(BigDecimal.ZERO, BigDecimal::add);
                     stat.setSum(sum);
                     return stat;
                 })
-                .sorted(Comparator.comparing(StatSrt::getSum))
-                .collect(Collectors.toList());
+                .sorted(Comparator.comparing(StatByString::getSum))
+                .toList();
 
             putMonthStat(monthStat, transactionList, monthStatistic);
         }
@@ -130,7 +117,7 @@ public class ExcelService {
         write(resultFileName, template);
     }
 
-    private void putMonthStat(Sheet sheet, List<Transaction> trnList, List<StatSrt> monthStatistic) {
+    private void putMonthStat(Sheet sheet, List<Transaction> trnList, List<StatByString> monthStatistic) {
         int rowIndex = 0;
         int colIndex = -1;
         for (Transaction trn : trnList) {
@@ -147,7 +134,7 @@ public class ExcelService {
 
         rowIndex = 0;
         colIndex += 1;
-        for (StatSrt statSrt : monthStatistic) {
+        for (StatByString statSrt : monthStatistic) {
             Row row = sheet.getRow(++rowIndex);
             int colIndexStat = colIndex;
 
@@ -165,23 +152,6 @@ public class ExcelService {
                 int year = calendar.get(Calendar.YEAR);
                 return String.format("%02d.%d", month, year);
             }));
-    }
-
-    public void createStat() {
-        String fileName = Const.WORK_DIR + "stat.xlsx";
-        {
-            File resFile = new File(fileName);
-            if (resFile.exists()) resFile.delete();
-        }
-
-        List<Transaction> transactions = readAndSortFilesTransactions();
-
-        List<Stat> stat = getStat(fillEmptyDays(transactions));
-
-        List<Stat> monthStat = getMonthStat(stat);
-        addSheetMonthStat(fileName, monthStat);
-        addSheetDaysStat(fileName, stat);
-        addSheetTrn(fileName, transactions);
     }
 
     private List<Transaction> readAndSortFilesTransactions() {
@@ -209,144 +179,6 @@ public class ExcelService {
         }
     }
 
-    private void addSheetMonthStat(String fileName, List<Stat> stats) {
-        Workbook workbook = readOrCreate(fileName);
-        putStat(workbook, "Статистика по місяцях", stats);
-        write(fileName, workbook);
-    }
-
-    private void addSheetDaysStat(String fileName, List<Stat> stats) {
-        Workbook workbook = readOrCreate(fileName);
-        putStat(workbook, "Статистика по дням", stats);
-        write(fileName, workbook);
-    }
-
-    private void addSheetTrn(String fileName, List<Transaction> transactions) {
-        Workbook workbook = readOrCreate(fileName);
-
-        removeSheet(workbook, "Виписка");
-        putTrnData(workbook, transactions);
-
-        write(fileName, workbook);
-    }
-
-    private void putStat(Sheet sheet, List<Stat> stats) {
-        int rowIndex = -1;
-        int colIndex = -1;
-        Row row = sheet.getRow(++rowIndex);
-        row.getCell(++colIndex).setCellValue("Date");
-        row.getCell(++colIndex).setCellValue("Sum");
-        for (Stat stat : stats) {
-            row = sheet.getRow(++rowIndex);
-            colIndex = -1;
-
-            Cell dateCell = row.getCell(++colIndex);
-            dateCell.setCellValue(stat.getDate());
-            row.getCell(++colIndex).setCellValue(stat.getSum().doubleValue());
-        }
-    }
-
-    private void putStat(Workbook workbook, String sheetName, List<Stat> stats) {
-        Sheet sheet = workbook.createSheet(sheetName);
-
-        CellStyle fullDateCellStyle = getTimeFormat(workbook, "dd.MM");
-        fullDateCellStyle.setAlignment(HorizontalAlignment.CENTER);
-
-        int rowIndex = 0;
-        int colIndex = 0;
-        Row row = sheet.createRow(rowIndex++);
-        row.createCell(colIndex++).setCellValue("Date");
-        row.createCell(colIndex++).setCellValue("Sum");
-        for (Stat stat : stats) {
-            row = sheet.createRow(rowIndex++);
-            colIndex = 0;
-
-            Cell dateCell = row.createCell(colIndex++);
-            dateCell.setCellValue(stat.getDate());
-            dateCell.setCellStyle(fullDateCellStyle);
-            row.createCell(colIndex++).setCellValue(stat.getSum().doubleValue());
-        }
-
-
-        sheet.setColumnWidth(0, (10+1)*256);
-    }
-
-    private void putTrnData(Workbook workbook, List<Transaction> transactions) {
-        Sheet sheet = workbook.createSheet("Виписка");
-
-        CellStyle fullDateCellStyle = getTimeFormat(workbook, "dd.MM.yyyy HH:mm");
-        fullDateCellStyle.setAlignment(HorizontalAlignment.CENTER);
-
-
-        int rowIndex = 0;
-        int colIndex = 0;
-        Row row = sheet.createRow(rowIndex++);
-        row.createCell(colIndex++).setCellValue("Повна дата");
-//        row.createCell(colIndex++).setCellValue("Дата");
-//        row.createCell(colIndex++).setCellValue("Час");
-        row.createCell(colIndex++).setCellValue("Категорія");
-        row.createCell(colIndex++).setCellValue("Опис");
-        row.createCell(colIndex++).setCellValue("Сума");
-
-        for (Transaction transaction : transactions) {
-            row = sheet.createRow(rowIndex++);
-            colIndex = 0;
-
-            Cell dateCell = row.createCell(colIndex++);
-            dateCell.setCellValue(transaction.getFullDate());
-            dateCell.setCellStyle(fullDateCellStyle);
-//            row.createCell(colIndex++).setCellValue(transaction.getDate());
-//            row.createCell(colIndex++).setCellValue(transaction.getTime());
-            row.createCell(colIndex++).setCellValue(transaction.getCategory());
-            row.createCell(colIndex++).setCellValue(transaction.getDscr());
-            row.createCell(colIndex++).setCellValue(transaction.getSum().doubleValue());
-            row.createCell(colIndex++).setCellFormula("MONTH(A" + rowIndex + ")");
-        }
-
-        sheet.setColumnHidden(4, true);
-        sheet.setColumnWidth(0, (16+1)*256);
-        sheet.setColumnWidth(1, (30+1)*256);
-        sheet.setColumnWidth(2, (40+1)*256);
-        sheet.setColumnWidth(3, (10+1)*256);
-        sheet.setAutoFilter(new CellRangeAddress(0, rowIndex-1,0, colIndex-1));
-    }
-
-    private List<Stat> getMonthStat(List<Stat> stat) {
-        List<Stat> result = new ArrayList<>();
-        BigDecimal monthSum = BigDecimal.ZERO;
-        Date month = stat.get(0).getDate();
-        for (Stat s : stat) {
-            if (!isSameMonth(month, s.getDate())) {
-                result.add(new Stat(month, monthSum));
-                monthSum = BigDecimal.ZERO;
-                month = s.getDate();
-            }
-            monthSum = monthSum.add(s.getSum());
-        }
-        result.add(new Stat(month, monthSum));
-        return result;
-    }
-
-    private List<Stat> getStat(List<Transaction> transactions) {
-        List<Stat> result = new ArrayList<>();
-
-        Map<Date, List<Transaction>> groupsByDays = transactions.stream().collect(Collectors.groupingBy(Transaction::getDate));
-        for (Date date : groupsByDays.keySet()) {
-            List<Transaction> day = groupsByDays.get(date);
-            BigDecimal sum = BigDecimal.ZERO;
-            for (Transaction trn : day) {
-                if (!ignore.contains(trn.getDscr())) {
-                    sum = sum.add(trn.getSum());
-                }
-            }
-            Stat stat = new Stat();
-            stat.setDate(date);
-            stat.setSum(sum);
-            result.add(stat);
-        }
-        return result.stream().sorted(Comparator.comparing(Stat::getDate)).collect(Collectors.toList());
-    }
-
     private List<Transaction> fillEmptyDays(List<Transaction> transactions) {
         Set<Date> dates = transactions.stream().map(Transaction::getDate).collect(Collectors.toSet());
         Date firstTrn = transactions.get(0).getDate();
@@ -354,12 +186,10 @@ public class ExcelService {
         Date tempDate = firstTrn;
         Calendar calendar = Calendar.getInstance();
         calendar.setTime(firstTrn);
-        int emptyDays = 0;
         while (tempDate.before(lastTrn)) {
             calendar.add(Calendar.DAY_OF_MONTH, 1);
             tempDate = calendar.getTime();
             if (!dates.contains(tempDate)) {
-                emptyDays++;
                 Transaction emptyDayTrn = new Transaction();
                 emptyDayTrn.setFullDate(tempDate);
                 emptyDayTrn.setDate(tempDate);
@@ -371,13 +201,4 @@ public class ExcelService {
         }
         return transactions.stream().sorted(Comparator.comparing(Transaction::getFullDate)).collect(Collectors.toList());
     }
-
-    private CellStyle getTimeFormat(Workbook workbook, String pattern) {
-        CellStyle fullDateCellStyle = workbook.createCellStyle();
-        CreationHelper createHelper = workbook.getCreationHelper();
-        short format = createHelper.createDataFormat().getFormat(pattern);
-        fullDateCellStyle.setDataFormat(format);
-        return fullDateCellStyle;
-    }
-
 }
